requirements.txt
dash==2.17.*
plotly>=5.22
waitress>=3.0
paho-mqtt>=1.6  # пригодится потом для MQTT-контроллера
